requirements.txt
streamlit>=1.35
numpy>=1.24
google-genai>=1.0
requests
